+++ main/src/observables/gpu_reductions.h
@@ -141,6 +141,6 @@ struct invalidAuxTEntry
  */
 template<class T, class Tc, class Th>
 extern std::tuple<std::vector<AuxT<T>>, std::vector<AuxT<T>>>
-localGrowthRateRTGpu(const int nAverage, size_t first, size_t last, Tc ymin, Tc ymax, const Th* h, const T* y,
+localGrowthRateRTGpu(const size_t nAverage, size_t first, size_t last, Tc ymin, Tc ymax, const Th* h, const T* y,
                      const Th* vy, const Th* markRamp);
 } // namespace sphexa
